requirements.txt
Django
django-jazzmin
django-import-export
django-shortuuidfield
Pillow